[__main__] Don't try to convert a VariableFontFace to a NormalFontFace twice
Move the conversion of variable font to ttc font in the same loop as where we request font to FontCollection

This allow the converted font (.ttc file) to be added into the FontCollection, so it doesn't collect the VariableFontFace twice. First, it collects the VariableFontFace that we convert to a NormalFontFace, so when we request the same variable font twice, it won't try to convert it to a ttc file twice.