Add --use-fonts-muxed-to-mkv
There is many problem by adding `--use-fonts-muxed-to-mkv` that I don't see how I could fix them.

Edge case:
The mkv contain a variable font with the family name "XXX".
The system contain a font (a normal one) with the same family name ("XXX").
Instead of using the mkv fonts (which is probably what the user want), FontCollector will delete the muxed font ("XXX") and mux the system font to the mkv.

Another edge case:
The mkv contain a variable font with the family name "XXX".
The system DOESN'T contain a font with the family name "XXX".
The option "--dont-convert-variable-to-collection" is NOT specified.
Instead of using the mkv font, it will convert the variable font "XXX" to a normal font. This may or may not be what the user want.

Another edge case:
The mkv contain a normal font with the family name "XXX".
The user specify `--additional-fonts` with the same family name ("XXX").
Instead of using the mkv fonts (which is probably what the user want), FontCollector will delete the muxed font ("XXX") and use the font provided in the `--additional-fonts` option.

PS: The mypy report an error, but in my opinion, this is a bug. I did open an issue about it: https://github.com/python/mypy/issues/20481